+++ src/core/se/jbee/inject/procedure/ProcedureMalfunction.java
@@ -3,11 +3,11 @@
  *			
  *  Licensed under the Apache License, Version 2.0, http://www.apache.org/licenses/LICENSE-2.0
  */
-package se.jbee.inject.service;
+package se.jbee.inject.procedure;
 
-public class ServiceMalfunction extends RuntimeException {
+public class ProcedureMalfunction extends RuntimeException {
 
-	public ServiceMalfunction(String message, Throwable cause) {
+	public ProcedureMalfunction(String message, Throwable cause) {
 		super(message, cause);
 	}