+++ src/test/se/jbee/inject/SuiteSilk.java
@@ -6,13 +6,13 @@ import org.junit.runners.Suite.SuiteClasses;
 
 import se.jbee.inject.bind.SuiteBind;
 import se.jbee.inject.container.SuitContainer;
-import se.jbee.inject.service.SuiteService;
+import se.jbee.inject.procedure.SuiteProcedure;
 
 @RunWith ( Suite.class )
 @SuiteClasses ( { TestName.class, TestType.class, TestPackages.class, TestMorePrecise.class,
 		TestTarget.class, TestDeclarationType.class,
 		// suits
-		SuitContainer.class, SuiteBind.class, SuiteService.class } )
+		SuitContainer.class, SuiteBind.class, SuiteProcedure.class } )
 public class SuiteSilk {
 	// all project tests
 }