+++ src/test/se/jbee/inject/procedure/ServiceInvocation.java
@@ -3,18 +3,19 @@
  *			
  *  Licensed under the Apache License, Version 2.0, http://www.apache.org/licenses/LICENSE-2.0
  */
-package se.jbee.inject.service;
+package se.jbee.inject.procedure;
 
 import se.jbee.inject.Type;
+import se.jbee.inject.procedure.Procedure;
 
 /**
- * Frames the invocation of {@link ServiceMethod}s with further functionality that can be executed
+ * Frames the invocation of {@link Procedure}s with further functionality that can be executed
  * {@link #before(Type, Object, Type)} or {@link #after(Type, Object, Type, Object, Object)} the invoked
- * {@linkplain ServiceMethod}. Thereby a state can be passed between these tow methods. The result
+ * {@linkplain Procedure}. Thereby a state can be passed between these tow methods. The result
  * of first will be passed to the second as an argument. This allows them stay stateless.
  * 
  * A {@link ServiceInvocation} intentionally doesn't give any control or access over/to the invoked
- * {@linkplain ServiceMethod} in order to be able to grant the same actual function invocation even
+ * {@linkplain Procedure} in order to be able to grant the same actual function invocation even
  * with faulty {@linkplain ServiceInvocation}s in place. That includes catching all exceptions
  * thrown in {@link #before(Type, Object, Type)} or {@link #after(Type, Object, Type, Object, Object)}.
  *